MAUI Classic Document: - Added a new page for Barcode Classic Component example. - Added a separate page for Custom implementation for Barcode Classic component - Commented the Document Scanner Classic Component as it is in progress.
