+++ MAUI/ClassicComponent.Maui/ClassicComponent.Maui.csproj
@@ -1,5 +1,5 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-    
+
     <PropertyGroup>
         <TargetFrameworks>net8.0-android;net8.0-ios</TargetFrameworks>
         <!-- Comment out the above and add the below to run with .net7 -->
@@ -9,14 +9,14 @@
         <SingleProject>true</SingleProject>
         <ImplicitUsings>enable</ImplicitUsings>
         <RootNamespace>ClassicComponent.Maui</RootNamespace>
-        
+
         <!-- Display name -->
         <ApplicationTitle>MAUI Classic Component</ApplicationTitle>
-        
+
         <!-- App Identifier -->
         <ApplicationId>io.scanbot.example.sdk.maui</ApplicationId>
         <ApplicationIdGuid>6742cc14-2939-49ab-8ffe-dd1b6567b00b</ApplicationIdGuid>
-        
+
         <!-- Versions -->
         <ApplicationDisplayVersion>1.0</ApplicationDisplayVersion>
         <ApplicationVersion>1</ApplicationVersion>
@@ -68,68 +68,69 @@
     </PropertyGroup>
 
     <!-- ===================== Application Resources =====================  -->
-    
+
     <ItemGroup>
         <!-- App Icon -->
-        <MauiIcon Include="Resources\AppIcon\appicon.svg" ForegroundFile="Resources\AppIcon\appiconfg.svg" Color="#512BD4" />
-        
+        <MauiIcon Include="Resources\AppIcon\appicon.svg" ForegroundFile="Resources\AppIcon\appiconfg.svg" Color="#512BD4"/>
+
         <!-- Splash Screen -->
-        <MauiSplashScreen Include="Resources\Splash\splash.svg" Color="#512BD4" BaseSize="128,128" />
-        
+        <MauiSplashScreen Include="Resources\Splash\splash.svg" Color="#512BD4" BaseSize="128,128"/>
+
         <!-- Images -->
-        <MauiImage Include="Resources\Images\*" />
-        <MauiImage Update="Resources\Images\dotnet_bot.svg" BaseSize="168,208" />
-        
+        <MauiImage Include="Resources\Images\*"/>
+        <MauiImage Update="Resources\Images\dotnet_bot.svg" BaseSize="168,208"/>
+
         <!-- Custom Fonts -->
-        <MauiFont Include="Resources\Fonts\*" />
-        
+        <MauiFont Include="Resources\Fonts\*"/>
+
         <!-- Raw Assets (also remove the "Resources\Raw" prefix) -->
-        <MauiAsset Include="Resources\Raw\**" LogicalName="%(RecursiveDir)%(Filename)%(Extension)" />
+        <MauiAsset Include="Resources\Raw\**" LogicalName="%(RecursiveDir)%(Filename)%(Extension)"/>
     </ItemGroup>
 
     <!-- ===================== Platform Specific files for Native Handler implementation =====================  -->
-    <ItemGroup  Condition="$(TargetFramework.Contains('-ios'))">
-        <Compile Remove="**\**\*.Android.cs" />
-        <None Include="**\**\*.Android.cs" Exclude="$(DefaultItemExcludes);$(DefaultExcludesInProjectFolder)" />
-        <Compile Update="Pages\DocumentScanner\ClassicDocumentScannerPage.xaml.cs">
-          <DependentUpon>ClassicDocumentScannerPage.xaml</DependentUpon>
-          <SubType>Code</SubType>
-        </Compile>
-        <Compile Update="Pages\DocumentScanner\DocumentScannerResultPage.xaml.cs">
-          <DependentUpon>DocumentScannerResultPage.xaml</DependentUpon>
-          <SubType>Code</SubType>
-        </Compile>
-        <Compile Update="Pages\BarcodeScanner\ClassicBarcodeScannerPage.xaml.cs">
-          <DependentUpon>ClassicBarcodeScannerPage.xaml</DependentUpon>
+    <ItemGroup Condition="$(TargetFramework.Contains('-ios'))">
+        <Compile Remove="**\**\*.Android.cs"/>
+        <None Include="**\**\*.Android.cs" Exclude="$(DefaultItemExcludes);$(DefaultExcludesInProjectFolder)"/>
+        <Compile Remove="Pages\DocumentScanner\ClassicDocumentScannerPage.xaml.cs" />
+        <Compile Update="Pages\HomePage.xaml.cs">
+          <DependentUpon>HomePage.xaml</DependentUpon>
           <SubType>Code</SubType>
         </Compile>
     </ItemGroup>
-    <ItemGroup  Condition="$(TargetFramework.Contains('-android'))">
-        <Compile Remove="**\**\*.iOS.cs" />
-        <None Include="**\**\*.iOS.cs" Exclude="$(DefaultItemExcludes);$(DefaultExcludesInProjectFolder)" />
+    <ItemGroup Condition="$(TargetFramework.Contains('-android'))">
+        <Compile Remove="**\**\*.iOS.cs"/>
+        <None Include="**\**\*.iOS.cs" Exclude="$(DefaultItemExcludes);$(DefaultExcludesInProjectFolder)"/>
     </ItemGroup>
-    
+
     <ItemGroup>
-      <Folder Include="Platforms\Android\Resources\" />
-      <Folder Include="Platforms\Android\Resources\layout\" />
-      <Folder Include="Platforms\Android\Resources\values\" />
-      <Folder Include="Platforms\iOS\Utils\" />
+        <Folder Include="Platforms\Android\Resources\"/>
+        <Folder Include="Platforms\Android\Resources\layout\"/>
+        <Folder Include="Platforms\Android\Resources\values\"/>
+        <Folder Include="Platforms\iOS\Utils\"/>
     </ItemGroup>
     <ItemGroup>
-      <None Remove="Platforms\iOS\Utils\" />
+        <None Remove="Platforms\iOS\Utils\"/>
     </ItemGroup>
     <ItemGroup>
         <PackageReference Include="ScanbotSDK.MAUI" Version="5.1.1" />
     </ItemGroup>
     <ItemGroup>
-      <MauiXaml Update="Pages\DocumentScanner\ClassicDocumentScannerPage.xaml">
-        <SubType>Designer</SubType>
-      </MauiXaml>
-      <MauiXaml Update="Pages\DocumentScanner\DocumentScannerResultPage.xaml">
-        <SubType>Designer</SubType>
-      </MauiXaml>
-      <MauiXaml Update="Pages\BarcodeScanner\ClassicBarcodeScannerPage.xaml">
-        <SubType>Designer</SubType>
-      </MauiXaml>
+        <MauiXaml Update="Pages\DocumentScanner\DocumentScannerResultPage.xaml">
+            <SubType>Designer</SubType>
+        </MauiXaml>
+        <MauiXaml Update="Pages\BarcodeScanner\ClassicBarcodeScannerPage.xaml">
+            <SubType>Designer</SubType>
+        </MauiXaml>
+        <Compile Update="Pages\DocumentScanner\DocumentScannerResultPage.xaml.cs">
+            <DependentUpon>DocumentScannerResultPage.xaml</DependentUpon>
+            <SubType>Code</SubType>
+        </Compile>
+        <Compile Update="Pages\BarcodeScanner\ClassicBarcodeScannerPage.xaml.cs">
+            <DependentUpon>ClassicBarcodeScannerPage.xaml</DependentUpon>
+            <SubType>Code</SubType>
+        </Compile>
+        <MauiXaml Update="Pages\HomePage.xaml">
+          <SubType>Designer</SubType>
+        </MauiXaml>
     </ItemGroup>
 </Project>
