+++ MAUI/ClassicComponent.Maui/Pages/BarcodeScanner/ClassicBarcodeScannerPage.xaml
@@ -3,61 +3,43 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              x:Class="ClassicComponent.Maui.ClassicBarcodeScannerPage"
              xmlns:ios="clr-namespace:Microsoft.Maui.Controls.PlatformConfiguration.iOSSpecific;assembly=Microsoft.Maui.Controls"
+             xmlns:classicComponent="clr-namespace:ScanbotSDK.MAUI.ClassicComponent;assembly=ScanbotSDK.Shared"
              xmlns:local="clr-namespace:ClassicComponent.Maui.CustomViews"
              ios:Page.UseSafeArea="true"
              BackgroundColor="#c8193c"
              xmlns:deviceDimensions="clr-namespace:Microsoft.Maui.Devices;assembly=Microsoft.Maui.Essentials">
+    <Grid RowSpacing="10">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="7*" />
+            <RowDefinition Height="3*" />
+        </Grid.RowDefinitions>
 
-    <StackLayout BackgroundColor="#e2e2e2" Spacing="0">
-        <Frame BackgroundColor="#c8193c" Padding="16" CornerRadius="0" HasShadow="False">
-            <StackLayout Orientation="Horizontal" HorizontalOptions="FillAndExpand">
-                <Label Text="Barcode Scanner Example" VerticalOptions="Center" HorizontalTextAlignment="Start" TextColor="White" FontSize="18"/>
-                <ImageButton x:Name="infoButton" Source="ic_info.png" HorizontalOptions="EndAndExpand" VerticalOptions="Center" BackgroundColor="Transparent" WidthRequest="24" HeightRequest="24"/>
+        <Grid Grid.Row="0">
+            <classicComponent:ScanbotClassicUIContainer>
+                <classicComponent:BarcodeScannerView
+                    x:Name="cameraView"
+                    OnBarcodeScanResult="CameraView_OnOnBarcodeScanResult" />
+            </classicComponent:ScanbotClassicUIContainer>
+        </Grid>
+        <ScrollView Grid.Row="1">
+            <StackLayout>
+                <Label
+                    x:Name="lblResultCaption"
+                    Text="Result"
+                    FontAttributes="Bold"
+                    FontSize="18"
+                    HorizontalTextAlignment="Center"
+                    VerticalTextAlignment="Center"
+                    VerticalOptions="Center"
+                    HorizontalOptions="Center" />
+                <Label
+                    x:Name="lblResult"
+                    Margin="15,10"
+                    HorizontalTextAlignment="Center"
+                    VerticalTextAlignment="Center"
+                    VerticalOptions="Center"
+                    HorizontalOptions="Center" />
             </StackLayout>
-        </Frame>
-
-        <!-- =============================================================================================================
-
-        Here we include our BarcodeCameraView. This view will be overriden by our native implementation because of
-        the custom renderers that we have created. See: BarcodeCameraView.cs and AndroidBarcodeCameraRenderer.cs.
-        Remember to define an xmlns attribute on the XAML root element to access your Views (in this case xmlns:local) -->
-        <StackLayout  BackgroundColor="#1f1f1f">
-            <local:BarcodeCameraView x:Name="cameraView" HeightRequest="300" WidthRequest="300" />
-        </StackLayout>
-
-        <!--==========================================================================================================-->
-
-        <StackLayout x:Name="buttonsLayout">
-            <BoxView Style="{StaticResource Separator}"/>
-
-            <Grid>
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="2*" />
-                    <ColumnDefinition Width="6*" />
-                    <ColumnDefinition Width="2*" />
-                </Grid.ColumnDefinitions>
-                <StackLayout Grid.Column="0"/>
-                <StackLayout Grid.Column="1" Orientation="Vertical" HorizontalOptions="CenterAndExpand" Margin="16, 32, 16, 32">
-                    <Button x:Name="scanButton" Text="START SCANNING" CornerRadius="10" Padding="24, 16" BackgroundColor="#c8193c" TextColor="White"/>
-                </StackLayout>
-                <Grid Grid.Column="2">
-                    <ImageButton x:Name="imgButtonFlash"
-                                 CornerRadius="20"
-                                 Clicked="FlashButtonClicked"
-                                 Source="ui_scanning_flash_on"
-                                 WidthRequest="40"
-                                 HeightRequest="40"
-                                 Padding="3"
-                                 BackgroundColor="Transparent"
-                                 HorizontalOptions="Center"
-                                 VerticalOptions="Center"/>
-                </Grid>
-            </Grid>
-            <BoxView Style="{StaticResource Separator}"/>
-        </StackLayout>
-        <StackLayout x:Name="resultsPreviewLayout" BackgroundColor="#d2d2d2" VerticalOptions="FillAndExpand" HorizontalOptions="FillAndExpand">
-            <Label Text="Results Preview (LIVE)" FontAttributes="Bold" Margin="16,16,16,16" FontSize="16" TextColor="Black"/>
-            <Label x:Name="resultsLabel" Margin="16,0,16,0" HorizontalTextAlignment="Start" TextColor="Black" FontSize="16"/>
-        </StackLayout>
-    </StackLayout>
+        </ScrollView>
+    </Grid>
 </ContentPage>
