+++ MAUI/ClassicComponent.Maui/Pages/HomePage.xaml
@@ -3,8 +3,7 @@
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              x:Class="ClassicComponent.Maui.HomePage"
-             xmlns:color="clr-namespace:ScanbotSDK.MAUI;assembly=ScanbotSDK.Shared"
-             Title="Scanbot Classic Components">
+             xmlns:color="clr-namespace:ScanbotSDK.MAUI;assembly=ScanbotSDK.Shared">
     <ContentPage.Content>
         <ScrollView>
             <StackLayout>
